+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   * Added the ability to change the forcing status with simple, short flags that can be used anywhere (Issue #7)
   * Support for referencing variables in other pipelines (Issue #15)
   * Support for omitting the .xp extension when using pipelines (Issue #18)
+  * Support for making unmarkable tasks
+  * Added support for task-level properties
 
 ### Changed